Restructure files to separate js and css from main key measures page file
